Move some function to utils folder
